+++ Build.sh
@@ -1,0 +1,2 @@
+pip install -U -r requirements/pypi.txt
+python setup.py sdist bdist_wheel