+++ aiomemcached/__init__.py
@@ -2,11 +2,21 @@
 
 Usage example::
 
-    import aiomcache
-    mc = aiomcache.Client("127.0.0.1", 11211, timeout=1, connect_timeout=5)
-    yield from mc.set("some_key", "Some value")
-    value = yield from mc.get("some_key")
-    yield from mc.delete("another_key")
+    import asyncio
+    import aiomemcached
+
+
+    async def example():
+        c = aiomemcached.Client()
+        await c.set(b"some_key", b"Some value")
+        value = await c.get(b"some_key")
+        print(value)
+        values = await c.multi_get(b"some_key", b"other_key")
+        print(values)
+        await c.delete(b"another_key")
+
+
+    asyncio.run(example())
 """
 
 from .client import Client